Move the LocaleCookieFactory to OpenConext\EngineBlockBundle\Http\Cookies namespace
