+++ src/OpenConext/EngineBlockBundle/Resources/config/services.yml
@@ -75,7 +75,7 @@ services:
             - '%kernel.default_locale%'
 
     engineblock.locale.cookie_factory:
-        class: OpenConext\EngineBlockBundle\Localization\LocaleCookieFactory
+        class: OpenConext\EngineBlockBundle\Http\Cookies\LocaleCookieFactory
         arguments:
             - '%locale_cookie_domain%'
             - '%locale_cookie_expiry%'
